Extract wallet icon tooltip to a separate component
